Differentiate signed and unsigned bundles about digests requirements
Unsigned bundles will primarily be used when developing, and to avoid
friction in tight dev/debug loops, we should have the digest field being
optional.

Also digests of container images in a thin bundle should not be
validated locally, but should match the digested reference of the image
manifest in the registry. This PR makes it explicit that thin-bundle
container image digests are not validated client-side, but are actually
used to create immutable references to registry images.

This second part is very important when we will work on sharing bundles
trough an image registry: we will need to reference all images in use in
a bundle trough their digested reference to ensure immutability, and to
protect us from deletion of referenced images.

Signed-off-by: Simon Ferquel <simon.ferquel@docker.com>