feat: define an extensions section of the manifest
This defines a section of the manifest where arbitrary extensions can be added.

The goal of this section is to provide a specific location for extension data to be defined such that we don't end up like `package.json`, where tons of ad hoc extensions have been added to the root document.