Add support for `bytes` parameter type
Some CNABs might have a need for parameters taking binary data as
values. An exemple of such a CNAB would be a generic Helm Chart deployer
declaring a HelmChart parameter that should be an actual Tarball with
the whole chart in it.
This commit adds support for this scenario and specify what defaultValue /
allowedValues should look like for parameters of type `bytes`.

Signed-off-by: Simon Ferquel <simon.ferquel@docker.com>