Add description and stateless fields to action
Custom actions lacked a description field. This fixed that.
Also introduces stateless custom actions to implement things like
`dry-run`, `help` etc.
The docker-app tool currently abuse the spec to do this kind of things
for the `inspect` and `render` actions, that are executed without
passing credentials, and are used to visualize the effect of parameter
values on the generated stack.

Signed-off-by: Simon Ferquel <simon.ferquel@docker.com>