remove base bundles from spec
Base bundles are pretty confusing and there is not
a good example of how to implement, use and distribute
base bundles and I'm not convinced this is a great route
to go anyway to in terms of composition and reuse of bundles.
I propose we take this section out completely to prevent
confusion and once we have a good story for composing/re-using
bundle functionality, then we can come back and add a recommendation
the 800 section.