Fix dependency example json (#216)
* Fix dependency example json

* When I went to add support for it in cnab-go, I realized that the
example json wasn't using array/object properly. This is what I get for
writing json embedded in markdown... 🤦‍♀️

* Use bool instead of bool string for prereleases

* Move dependencies into a custom extension

Since this isn't part of the core spec, it can't live at the root
because that messes up the json schema valiation. So it will live under
the custom extensions, custom.

* Add a json schema document
* Add a validation test for the schema doc

* Update dependencies spec id to match where it can be downloaded

I'll submit a follow-on PR to the cnab.io site to make it available once it's merged

* Define a key for each dependency

* Mark 'custom.dependencies' as reserved