Split claim into execution inputs and results
Any fields that are modified based on the bundle execution: outputs,
modified have moved to claim.results. The action, e.g. "install", has
always been an input and has moved out of result to the top level claim.

Now claim (with the exception of its child result) does not change after
creation. A tool could create a digest from it if needed.

Also results.outputs has been modified to only contain the output names
instead of the entire output contents, which can be prohibitively large.
The tool can request the outputs at the end of the operation to persist
them separate from the claim.

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>