Clarify pending status, introduce running
The current definition for the pending (previously underway) status is
difficult to understand when it is acceptable to use and what it really
means about the state of the installation. It's described as both a long
term state and temporary. It is also described as "to be avoided".

We would like to redefine the state to make it more clear on what the
state means, and to make it acceptable for tools to use. Pending now
means that the operation has been requested (most likely queued) but has
not begun executing. Tools that are immediately executing commands as
they are issued would most likely choose to not use this state. However
solutions that accept bundle operations from many users at once, for
example a cnab operator, may choose to use this state.

This also introduces a new state: running. Tools may use this state to
indicate that a bundle is currently executing an operation. This can be
used by tools to provide operational insight into the activity of the
system. For example, when an upgrade takes 30 minutes and people don't
know that it's already running, it could easily be initiated more than
once by multiple people on the same installation. Allowing tools to make
the current status of the installation visible, helps avoid this
situation.

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>