Have created and modified for claim and result
When claim and result were on the same document, we briefly considered
splitting created/modified between the top level claim and the result.
Undo that change and have a created/modified pair for both claim and
result.

Tools may choose to treat claim as immutable, but that isn't defined in
the spec.

Mark both created/modified as required, matching the schema and how they
are used in cnab-go.

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>