Clarify how to use claims and results
* Installation refers to the installation of the bundle
* Claim records the execution of an action against an installation
* Claims have an id that results reference
* An installation has _many_ claims with these new definitions.
* Claims and results are immutable. Each new action against an
installation generates a new claim. Each new status transition of an
operation could yield a new result (depending on the implementation).
* Clarify that the last modifying claim is retained (same as before)
unless the implemenation wants more history. The last result is retained
for a retained claim (again unless an implemenation wants more history).

Depending on your previous interpretation of the spec, this may change
what claim.created means to you. The meaning has hopefully been
clarified to the created date of the claim (see updated definition
above), not the installation.

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>