Split claim and claim result into separate documents (#349)
* Split claim and result into separate documents

Extract claim.result into a separate document and schema.

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>

* Claim result message is optional

Update schema to match the spec and cnab-go, the result's message is
optional.

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>

* Have created and modified for claim and result

When claim and result were on the same document, we briefly considered
splitting created/modified between the top level claim and the result.
Undo that change and have a created/modified pair for both claim and
result.

Tools may choose to treat claim as immutable, but that isn't defined in
the spec.

Mark both created/modified as required, matching the schema and how they
are used in cnab-go.

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>

* Incorporate review feedback

* Clarify claim created doesn't change with new revisions
* Clarify that outputs does not contain the output value anymore, but the output metadata
* Fix result schema to make claimRef.installation and revision required

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>

* Clarify how to use claims and results

* Installation refers to the installation of the bundle
* Claim records the execution of an action against an installation
* Claims have an id that results reference
* An installation has _many_ claims with these new definitions.
* Claims and results are immutable. Each new action against an
installation generates a new claim. Each new status transition of an
operation could yield a new result (depending on the implementation).
* Clarify that the last modifying claim is retained (same as before)
unless the implemenation wants more history. The last result is retained
for a retained claim (again unless an implemenation wants more history).

Depending on your previous interpretation of the spec, this may change
what claim.created means to you. The meaning has hopefully been
clarified to the created date of the claim (see updated definition
above), not the installation.

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>

* Add contentDigest to claim output metadata

The optional contentDigest can be added by the runtime so that tooling
can validate the contents of the output requested from the runtime, or
at any later date if it is stored.

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>

* Apply suggestions from code review

Co-Authored-By: Jacob LeGrone <jlegrone@users.noreply.github.com>
Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>

* Replace an ulid with a ulid

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>

Co-authored-by: Jacob LeGrone <jlegrone@users.noreply.github.com>