Add parameter sources custom extension
This custom extension defines how a runtime may connect outputs back to
parameters. It is terribly useful in reducing user input but is also
necessary now that we don't store the entire contents of outputs in the claim
anymore. It gives authors a way to signal to runtimes when an output
matters and when it should be injected back into a bundle and on which
actions.

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>