Add parameter sources custom extension (#365)
* Add parameter sources custom extension

This custom extension defines how a runtime may connect outputs back to
parameters. It is terribly useful in reducing user input but is also
necessary now that we don't store the entire contents of outputs in the claim
anymore. It gives authors a way to signal to runtimes when an output
matters and when it should be injected back into a bundle and on which
actions.

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>

* Fix typo

Co-authored-by: Vaughn Dice <vadice@microsoft.com>
Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>

* Incorporate feedback

* Clarify responsibility of invocation image and runtime
* Allow for multiple sources and priorities
* Define well known sources

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>

* Add parameter sources schema and example

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>

Co-authored-by: Vaughn Dice <vadice@microsoft.com>