Support namespace and labels
* Namespaces allow us to scope documents to a particular namespace or leave empty to define it in the global namespace.
* Labels let us define key/value pairs on data stored that we can use to query later.
* Add labels to bundles, credential sets, installations, claims, results, and outputs
* Add namespaces to credential sets, installations, claims, results, and outputs.

Bundles already had the concept of namespacing via using the format com.example.mybundle. I did not want to use this format for the other documents and opted to use a separate field to preserve existing behavior. People could have named their other documents with periods not intending for it to be interpreted as namespaces.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>