Use status canceled instead of cancelled
The spec originally used a variation of the word "canceled" and it was
spelled "cancelled". However all of the CNAB implementations (duffle,
porter, Data Dog) all used "canceled".

We have agreed to use the term that is already in everyone's data, and
this updates the spec to use the term "canceled".

Fixes #412

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>